+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0 - 2026-02-02
+## 1.1.0
+
+- Remove `showStickyHeader` properties
+- Fix the abnormal alphabetic index problem when `sticky` is `false`
+
+## 1.0.0
 
 - Initial release with sticky headers and index bar / 初始版本，支持粘性分组头与字母索引条